+++ xds/googledirectpath/googlec2p.go
@@ -182,13 +182,7 @@ func newNodeConfig(zone string, ipv6Capable bool) map[string]any {
 		"id":       fmt.Sprintf("C2P-%d", randInt()),
 		"locality": map[string]any{"zone": zone},
 	}
-	if envconfig.NewPickFirstEnabled {
-		// Enable dualstack endpoints in TD.
-		// TODO(https://github.com/grpc/grpc-go/issues/8561): remove IPv6 metadata server queries entirely after old pick first is removed.
-		ipv6Capable = true
-	} else {
-		logger.Infof("GRPC_EXPERIMENTAL_ENABLE_NEW_PICK_FIRST is disabled, setting ipv6Capable node metadata based on metadata server query")
-	}
+	// Enable dualstack endpoints in TD.
 	if ipv6Capable {
 		node["metadata"] = map[string]any{ipv6CapableMetadataName: true}
 	}
